change of circle algorithm
it's a bit of a hack, and it's probably slightly slower than before, but it no
longer requires precalculating/storing a number of values, with n=radius. the
previous approach was worrying because big circles required a lot of scratch
storage, and while I might've been able to limit the amount of storage required
to the image height instead of the circle radius, I'm not sure I could've
easily skipped the steps, because DDA is incremental. the new approach
shouldn't have degenerate cases like these.